Media system fix

+++ App/Controllers/SETUP/SetupController.php
@@ -12,40 +12,33 @@ class SetupController extends Controller
     public function DatabaseSetup(Request $request)
     {
         try {
-            $db["host"] = post("host");
-            $db["user"] = post("user");
-            $db["password"] = post("password");
-            $db["database"] = post("database");
-            $db["prefix"] = post("prefix");
 
-            $db["host"] = $db["host"] ? $db["host"] : config("DB_HOST");
-            $db["user"] = $db["user"] ? $db["user"] : config("DB_USER");
-            $db["password"] = $db["password"] ? $db["password"] : config("DB_PASSWORD");
-            $db["database"] = $db["database"] ? $db["database"] : config("DB_NAME");
-            $db["prefix"] = $db["prefix"] ? $db["prefix"] : config("DB_PREFIX");
+            $entityBody = json_decode($request->getContent(), true);
+            $db["host"] = isset($entityBody["host"]) ? dataClear($entityBody["host"]) : config("DB_HOST");
+            $db["user"] = isset($entityBody["user"]) ? dataClear($entityBody["user"]) : config("DB_USER");
+            $db["password"] = isset($entityBody["password"]) ? dataClear($entityBody["password"]) : config("DB_PASSWORD");
+            $db["database"] = isset($entityBody["database"]) ? dataClear($entityBody["database"]) : config("DB_NAME");
+            $db["prefix"] = isset($entityBody["prefix"]) ? dataClear($entityBody["prefix"]) : config("DB_PREFIX");
 
 
-            $dbEnv = "<?php 
-    \$dbENV = [
-        'DB_HOST' => '" . $db["host"] . "',
-        'DB_NAME' => '" . $db["database"] . "',
-        'DB_USER' => '" . $db["user"] . "',
-        'DB_PASSWORD' => '" . $db["password"] . "',
-        'DB_PREFIX' => '" . $db["prefix"] . "',
-    ];
-    ?>";
+            $dbEnv = "<?php  \$dbENV = [
+    'DB_HOST' => '" . $db["host"] . "',
+    'DB_NAME' => '" . $db["database"] . "',
+    'DB_USER' => '" . $db["user"] . "',
+    'DB_PASSWORD' => '" . $db["password"] . "',
+    'DB_PREFIX' => '" . $db["prefix"] . "',
+]; ?>";
 
-            $write = file_put_contents(__DIR__ . '/../../ENV/db.env.php', $dbEnv);
+            $dbFilePath = __DIR__ . '/../../../ENV/db.env.php';
+
+            $write = file_put_contents($dbFilePath, $dbEnv);
             if ($write) {
                 response([
                     'status' => true,
                     'message' => 'Database Setup Success'
                 ]);
             } else {
-                response([
-                    'status' => true,
-                    'message' => 'Database Setup Failed'
-                ], 500);
+                throw new \Exception("Database Setup Failed");
             }
         } catch (\Exception $e) {
             response([
@@ -58,43 +51,51 @@ class SetupController extends Controller
 
     public function SuperAdminSetup(Request $request)
     {
-        $headers = $request->headers->all();
-
         try {
-            $adminName = post("name");
-            $adminPassword = post("password");
 
-            $adminName = $adminName ? $adminName : "admin";
-            $adminPassword = $adminPassword ? $adminPassword : "admin";
+            $entityBody = json_decode($request->getContent(), true);
+
+            $adminName = isset($entityBody["name"]) ? dataClear($entityBody["name"]) : "admin";
+            $adminPassword = isset($entityBody["password"]) ? dataClear($entityBody["password"]) : "admin";
             $adminPassword = passwordHash($adminPassword);
             $adminTableName = "super_admin";
 
             $adminExistSql = sql("SELECT * FROM information_schema.tables WHERE table_schema = '" . config("DB_NAME") . "' AND table_name = '$adminTableName' LIMIT 1");
 
-            if ($adminExistSql) {
+
+            $token = md5(uniqid());
+
+            $insertSQL = "INSERT INTO $adminTableName SET superadmin_name = '$adminName', superadmin_password= '$adminPassword', superadmin_token = '$token' ";
+
+
+            if ($adminExistSql && count($adminExistSql) > 0) {
                 // truncate table
                 sql("TRUNCATE TABLE $adminTableName");
-                sql("INSERT INTO $adminTableName SET superadmin_name = '$adminName', superadmin_password= '$adminPassword' ");
-                controllerResponse($headers, "Admin Setup Success");
+                sql($insertSQL);
+
+                response([
+                    'status' => true,
+                    'message' => 'Super Admin Setup Success'
+                ]);
             } else {
-                sql("CREATE TABLE `$adminTableName` (
-                `id` int(11) NOT NULL,
-                `superadmin_name` varchar(100) NOT NULL,
-                `superadmin_password` text NOT NULL,
-                `superadmin_status` tinyint(1) NOT NULL DEFAULT '1',
-                `superadmin_token` text NOT NULL,                
-              ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;");
 
-                $token = md5(uniqid());
+                $createSQL = "CREATE TABLE IF NOT EXISTS `super_admin`(
+                        `id` INT(11) NOT NULL AUTO_INCREMENT,
+                        `superadmin_name` VARCHAR(100) NOT NULL,
+                        `superadmin_password` TEXT NOT NULL,
+                        `superadmin_status` TINYINT(1) NOT NULL DEFAULT '1',
+                        `superadmin_token` TEXT NOT NULL,
+                        PRIMARY KEY(`id`)
+                    ) CHARSET = utf8 AUTO_INCREMENT = 1;";
 
-                sql("INSERT INTO $adminTableName SET superadmin_name = '$adminName', superadmin_password= '$adminPassword', superadmin_token = '$token' ");
+
+                sql($createSQL);
+                sql($insertSQL);
 
                 response([
                     'status' => true,
                     'message' => 'Admin Setup Success'
                 ]);
-
-
             }
         } catch (\Exception $e) {
             response([
@@ -113,8 +114,6 @@ class SetupController extends Controller
 
     public function FinishSetup(Request $request)
     {
-
-        $headers = $request->headers->all();
         try {
 
             $mediaLibraryCreateTableSql = "CREATE TABLE if NOT EXISTS media_library(
@@ -123,7 +122,7 @@ class SetupController extends Controller
                     `media_name` VARCHAR(400) NOT NULL,
                     `media_type` VARCHAR(100) NOT NULL,
                     `media_ext` VARCHAR(100) NOT NULL,
-                    `media_path` TEXT NOT NULL,
+                    `media_paths` TEXT NOT NULL,
                     `media_size` VARCHAR(100) NOT NULL,
                     `media_status` TINYINT(1) NOT NULL DEFAULT '1',
                     `media_created_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP(),
@@ -137,7 +136,7 @@ class SetupController extends Controller
 
             response([
                 'status' => true,
-                'message' => 'Assets Setup Success'
+                'message' => 'Media Library Setup Success'
             ]);
